requirements.txt
requests>=2.31
python-dotenv>=1.0
tabulate>=0.9
slack-bolt>=1.18
anthropic>=0.30
